refactor: move character sprites into character
