+++ frontend/src/game/character/sprites/idle/index.ts
