+++ frontend/src/game/character/sprites/walk/index.png
