+++ frontend/src/game/character/states.ts
@@ -2,9 +2,9 @@ import { StateConfig } from '../entity';
 import { Rectangle } from '../shapes';
 import Box from '../box';
 
-import walk from '../sprites/character/walk';
-import run from '../sprites/character/run';
-import idle from '../sprites/character/idle';
+import walk from './sprites/walk';
+import run from './sprites/run';
+import idle from './sprites/idle';
 
 enum CharacterState {
   Idle,
